requirements.txt
# Blockchain Forensics Tool - Dependencies
# M.Sc. Forensic Science (Cyber Forensics) - 9th Semester
# Author: Saransh Jhariya | 102FSBSMS2122012

# Core HTTP and API
requests>=2.31.0
web3>=6.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
jsonschema>=4.17.0

# Graph Analysis
networkx>=3.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
pyvis>=0.3.2
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
